chore: cleans up setup action
