+++ actions/setup/package.json
@@ -1,9 +1,12 @@
 {
   "name": "setup",
-  "version": "1.0.0",
-  "description": "",
+  "version": "0.1.0",
+  "description": "Setup the Omashu CLI on the runner",
   "main": "index.js",
   "scripts": {
+    "format": "prettier --write '**/*.ts'",
+    "lint": "eslint ./src",
+    "package": "ncc build src/index.ts --license licenses.txt",
     "test": "jest"
   },
   "keywords": [],
@@ -17,9 +20,51 @@
   "devDependencies": {
     "@types/jest": "^29.5.4",
     "@types/node": "^20.5.3",
+    "@typescript-eslint/eslint-plugin": "^6.4.1",
+    "@typescript-eslint/parser": "^6.4.1",
+    "@vercel/ncc": "^0.36.1",
+    "eslint": "^8.47.0",
+    "eslint-plugin-github": "^4.9.2",
+    "eslint-plugin-jest": "^27.2.3",
     "jest": "^29.6.3",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
     "typescript": "^5.1.6"
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "verbose": true,
+    "clearMocks": true,
+    "testEnvironment": "node",
+    "testMatch": [
+      "**/*.test.ts"
+    ],
+    "testPathIgnorePatterns": [
+      "/node_modules/",
+      "/dist/"
+    ],
+    "coverageReporters": [
+      "text"
+    ],
+    "collectCoverage": true,
+    "collectCoverageFrom": [
+      "./src/**"
+    ]
+  },
+  "prettier": {
+    "printWidth": 80,
+    "tabWidth": 2,
+    "useTabs": false,
+    "semi": false,
+    "singleQuote": true,
+    "quoteProps": "as-needed",
+    "jsxSingleQuote": false,
+    "trailingComma": "none",
+    "bracketSpacing": true,
+    "bracketSameLine": true,
+    "arrowParens": "avoid",
+    "proseWrap": "always",
+    "htmlWhitespaceSensitivity": "css",
+    "endOfLine": "lf"
   }
 }
